removed merge cruft
