+++ services/xap9.1-lite/README.md
@@ -1,0 +1,28 @@
+#XAP9.1-lite recipe notes:
+
+**Status**: 
+**Description**: XAP 9.1 Lite Version
+**Maintainer**: Cloudify
+**Contributors**: [dfilppi] https://github.com/dfilppi
+**Homepage**: [http://www.cloudifysource.org](http://www.cloudifysource.org)
+**Demo**:
+**License**: Apache 2.0
+**Linux* sudoer permissions**: Required
+**Windows* admin permissions**: NA
+**Release Date**: Work In Progress
+
+Tested on:
+-----------
+
+* <strong>EC2</strong>
+ *<strong>Amazon Linux</strong> imageId "us-east-1/ami-76f0061f", hardwareId "m1.small", locationId "us-east-1"
+* <strong>HPCS/Openstack</strong>
+ *<strong>Centos 5.6</strong> imageId "20111207", hardwareId "standard.medium", locationId "110"
+
+Synopsis:
+--------
+
+
+Known limitations:
+---------------
+
