+++ services/xap9.1-lite/runxap.bat
@@ -1,0 +1,35 @@
+SET CLASSPATH=
+SET bootclasspath=
+SET CLASS=
+SET COMMAND=
+SET CP=
+SET EXT_JARS=
+SET EXT_JAVA_OPTIONS=
+SET GS_JARS=
+SET GS_LIB=
+SET GS_LOGGING_CONFIG_FILE=
+SET GS_LOGGING_CONFIG_FILE_PROP=
+SET JAVA_OPTIONS=
+SET JSHOMEDIR=
+SET LOOKUPGROUPS=
+SET SIGAR_JARS=
+SET PRE_CLASSPATH=
+SET POST_CLASSPATH=
+SET AGENT_ID=
+SET GSA_SERVICE_ID=
+SET ENABLE_DYNAMIC_LOCATORS=
+SET RMI_OPTIONS=
+SET LOOKUP_LOCATORS_PROP=
+SET LOOKUP_GROUPS_PROP=
+SET command_line=
+SET startParm=
+SET launchTarget=
+SET options=
+
+SET root=install
+
+start /b %root%\gigaspaces-xap-9.1-lite\gigaspaces-xap-premium-9.1.0-ga\bin\gsm.bat
+start /b %root%\gigaspaces-xap-9.1-lite\gigaspaces-xap-premium-9.1.0-ga\bin\gsc.bat
+
+cmd /c %root%\gigaspaces-xap-9.1-lite\gigaspaces-xap-premium-9.1.0-ga\bin\gs-webui.bat
+
