+++ services/xap9.1-lite/xap-service.properties
@@ -1,0 +1,10 @@
+version = "9.1"
+name = "gigaspaces-xap-${version}-lite"
+zipName = "${name}.zip"
+installDir = "install"
+downloadPath = "https://s3.amazonaws.com/xap_9.1_lite/gigaspaces-xap-9.1-lite.zip"
+uiPort=8999
+
+xapDir="${installDir}/${name}/gigaspaces-xap-premium-9.1.0-ga"
+binDir="${xapDir}/bin"
+gscSize="2G"
